requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langchain-huggingface
langchain-text-splitters
faiss-cpu
python-dotenv
sentence-transformers